requirements.txt
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3

celery==5.6.2
redis==5.0.4
django-celery-beat==2.6.0

python-dateutil==2.9.0.post0
